Moved login stuff to separate login controller, removed redundant create method from early testing
